+++ src/main/java/com/iemr/mmu/repo/masterrepo/nurse/FamilyMemberMasterRepo.java
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Repository;
 
 import com.iemr.mmu.data.masterdata.nurse.FamilyMemberType;
 @Repository
-public interface FamilyMemberMasterRepo extends CrudRepository<FamilyMemberType, Long>{
+public interface FamilyMemberMasterRepo extends CrudRepository<FamilyMemberType, Short>{
 
 	@Query("select benRelationshipID, benRelationshipType, gender  from FamilyMemberType where deleted = false and benRelationshipID>3 order by benRelationshipType ")
 	public ArrayList<Object[]> getFamilyMemberTypeMaster();
