+++ src/main/java/com/iemr/mmu/repo/masterrepo/nurse/VisitCategoryMasterRepo.java
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Repository;
 
 import com.iemr.mmu.data.masterdata.nurse.VisitCategory;
 @Repository
-public interface VisitCategoryMasterRepo extends CrudRepository<VisitCategory, Long>{
+public interface VisitCategoryMasterRepo extends CrudRepository<VisitCategory, Short>{
 
 	@Query("select visitCategoryID, visitCategory from VisitCategory where deleted = false order by visitCategory ")
 	public ArrayList<Object[]> getVisitCategoryMaster();
