+++ src/main/java/com/iemr/mmu/repo/nurse/pnc/PNCCareRepo.java
@@ -35,7 +35,7 @@ import com.iemr.mmu.data.pnc.PNCCare;
 import jakarta.transaction.Transactional;
 
 @Repository
-public interface PNCCareRepo extends CrudRepository<PNCCare, Integer>{
+public interface PNCCareRepo extends CrudRepository<PNCCare, Long>{
 	
 	@Query(" SELECT ID, beneficiaryRegID, benVisitID, providerServiceMapID, visitNo, deliveryTypeID, deliveryType, deliveryPlaceID, "
 			+ "deliveryPlace, otherDeliveryPlace, dateOfDelivery, deliveryComplicationID, deliveryComplication, otherDeliveryComplication, "
